chore: moved around test cases and error message fix
