+++ unified_planning/test/examples/interpreted_functions_examples.py
@@ -347,4 +347,230 @@ def get_example_problems():
     )
     problems["treasure_hunting_robot_simple"] = ifproblem
 
+    def i_f_simple_bool(inputone, inputtwo):
+        return (inputone * inputtwo) == 60
+
+    problem = Problem("interpreted_functions_in_conditions")
+
+    signatureConditionF = OrderedDict()
+    signatureConditionF["inputone"] = IntType()
+    signatureConditionF["inputtwo"] = IntType()
+    funx = InterpretedFunction("funx", BoolType(), signatureConditionF, i_f_simple_bool)
+    end_goal = Fluent("end_goal")
+    ione = Fluent("ione", IntType(0, 20))
+    itwo = Fluent("itwo", IntType(0, 20))
+    instant_action_i_f_condition = InstantaneousAction("instant_action_i_f_condition")
+    increase_val = InstantaneousAction("increase_val")
+    increase_val.add_effect(itwo, Plus(itwo, 2))
+    instant_action_i_f_condition.add_precondition((funx(ione, itwo)))
+    instant_action_i_f_condition.add_effect(end_goal, True)
+    problem.add_fluent(end_goal)
+    problem.add_fluent(ione)
+    problem.add_fluent(itwo)
+    problem.add_action(increase_val)
+    problem.add_action(instant_action_i_f_condition)
+    problem.set_initial_value(end_goal, False)
+    problem.set_initial_value(ione, 12)
+    problem.set_initial_value(itwo, 3)
+    problem.add_goal(end_goal)
+    ifproblem = TestCase(
+        problem=problem,
+        solvable=True,
+        valid_plans=[
+            up.plans.SequentialPlan([increase_val(), instant_action_i_f_condition()]),
+        ],
+    )
+    problems["interpreted_functions_in_conditions"] = ifproblem
+
+    # interpreted functions in instaneous action precondition unsolvable even ignoring interpreted functions
+    problem = Problem("interpreted_functions_in_conditions_always_impossible")
+
+    signatureConditionF = OrderedDict()
+    signatureConditionF["inputone"] = IntType()
+    signatureConditionF["inputtwo"] = IntType()
+    funx = InterpretedFunction("funx", BoolType(), signatureConditionF, i_f_simple_bool)
+    end_goal = Fluent("end_goal")
+    ione = Fluent("ione", IntType(0, 20))
+    itwo = Fluent("itwo", IntType(0, 20))
+    instant_action_i_f_condition = InstantaneousAction("instant_action_i_f_condition")
+    instant_action_i_f_condition.add_precondition(And((funx(ione, itwo)), GE(ione, 19)))
+    instant_action_i_f_condition.add_effect(end_goal, True)
+    problem.add_fluent(end_goal)
+    problem.add_fluent(ione)
+    problem.add_fluent(itwo)
+    problem.add_action(increase_val)
+    problem.add_action(instant_action_i_f_condition)
+    problem.set_initial_value(end_goal, False)
+    problem.set_initial_value(ione, 12)
+    problem.set_initial_value(itwo, 3)
+    problem.add_goal(end_goal)
+    ifproblem = TestCase(
+        problem=problem,
+        solvable=False,
+    )
+    problems["interpreted_functions_in_conditions_always_impossible"] = ifproblem
+
+    # interpreted functions in durative action condition
+    funx = InterpretedFunction("funx", BoolType(), signatureConditionF, i_f_simple_bool)
+
+    end_goal = Fluent("end_goal")
+    ione = Fluent("ione", IntType(0, 20))
+    itwo = Fluent("itwo", IntType(0, 20))
+
+    durative_action_i_f_condition = DurativeAction("durative_action_i_f_condition")
+
+    durative_action_i_f_condition.add_condition(
+        EndTiming(), And(GE(ione, 10), Not(funx(itwo, itwo)))
+    )
+    durative_action_i_f_condition.add_condition(StartTiming(), (funx(ione, itwo)))
+    durative_action_i_f_condition.add_condition(
+        StartTiming(), Not(And(GE(ione, 15), LE(itwo, 5)))
+    )
+    durative_action_i_f_condition.add_effect(EndTiming(), end_goal, True)
+    durative_action_i_f_condition.set_fixed_duration(5)
+
+    problem = Problem("interpreted_functions_in_durative_conditions")
+    problem.add_fluent(end_goal)
+    problem.add_fluent(ione)
+    problem.add_fluent(itwo)
+    problem.add_action(durative_action_i_f_condition)
+    problem.set_initial_value(end_goal, False)
+    problem.set_initial_value(ione, 12)
+    problem.set_initial_value(itwo, 5)
+    problem.add_goal(end_goal)
+
+    ifproblem = TestCase(
+        problem=problem,
+        solvable=True,
+        valid_plans=[
+            up.plans.SequentialPlan([durative_action_i_f_condition()]),
+        ],
+        invalid_plans=[
+            up.plans.SequentialPlan([]),
+        ],
+    )
+    problems["interpreted_functions_in_durative_conditions"] = ifproblem
+
+    # interpreted functions in bool assignment
+
+    funx = InterpretedFunction("funx", BoolType(), signatureConditionF, i_f_simple_bool)
+
+    ione = Fluent("ione", IntType(0, 20))
+    itwo = Fluent("itwo", IntType(0, 20))
+    end_goal = Fluent("end_goal")
+
+    apply_i_f_assignment = InstantaneousAction("apply_i_f_assignment")
+    apply_i_f_assignment.add_effect(end_goal, funx(ione, itwo))
+
+    increase_val = InstantaneousAction("increase_val")
+    increase_val.add_effect(ione, Plus(ione, 2))
+
+    problem = Problem("interpreted_functions_in_boolean_assignment")
+    problem.add_fluent(ione)
+    problem.add_fluent(itwo)
+    problem.add_fluent(end_goal)
+
+    problem.add_action(apply_i_f_assignment)
+    problem.add_action(increase_val)
+    problem.set_initial_value(ione, 3)
+    problem.set_initial_value(itwo, 12)
+    problem.set_initial_value(end_goal, False)
+    problem.add_goal(end_goal)
+
+    ifproblem = TestCase(
+        problem=problem,
+        solvable=True,
+        valid_plans=[
+            up.plans.SequentialPlan([increase_val(), apply_i_f_assignment()]),
+        ],
+        invalid_plans=[
+            up.plans.SequentialPlan([apply_i_f_assignment(), increase_val()]),
+        ],
+    )
+    problems["interpreted_functions_in_boolean_assignment"] = ifproblem
+
+    # interpreted functions in numeric assignment
+
+    def i_f_simple_int(inputone, inputtwo):
+
+        return inputtwo * inputone
+
+    i_f_simple_int_if = InterpretedFunction(
+        "i_f_simple_int", IntType(), signatureConditionF, i_f_simple_int
+    )
+
+    ione = Fluent("ione", IntType(0, 20))
+    itwo = Fluent("itwo", IntType(0, 20))
+    end_goal = Fluent("end_goal", IntType(0, 20))
+
+    apply_i_f_assignment = InstantaneousAction("apply_i_f_assignment")
+    apply_i_f_assignment.add_effect(end_goal, i_f_simple_int_if(ione, itwo))
+
+    problem = Problem("interpreted_functions_in_numeric_assignment")
+    problem.add_fluent(ione)
+    problem.add_fluent(itwo)
+    problem.add_fluent(end_goal)
+
+    problem.add_action(apply_i_f_assignment)
+    problem.set_initial_value(ione, 3)
+    problem.set_initial_value(itwo, 5)
+    problem.set_initial_value(end_goal, 0)
+    problem.add_goal(GE(end_goal, 2))
+
+    ifproblem = TestCase(
+        problem=problem,
+        solvable=True,
+        valid_plans=[up.plans.SequentialPlan([apply_i_f_assignment()])],
+        invalid_plans=[
+            up.plans.SequentialPlan([]),
+        ],
+    )
+    problems["interpreted_functions_in_numeric_assignment"] = ifproblem
+
+    # interpreted functions minimal chain of assignments
+
+    def assign_plus_two(in_val):
+        out_val = in_val + 2
+        return out_val
+
+    signature_if_assign = OrderedDict()
+    signature_if_assign["i_one"] = IntType()
+    if_assign_plus_two = InterpretedFunction(
+        "if_assing", IntType(), signature_if_assign, assign_plus_two
+    )
+
+    start_value = Fluent("start_value", IntType())
+    mid_value = Fluent("mid_value", IntType())
+
+    g = Fluent("g", IntType())
+    start_action = InstantaneousAction("start_action")
+    start_action.add_effect(start_value, if_assign_plus_two(start_value))
+    mid_action = InstantaneousAction("mid_action")
+    mid_action.add_effect(mid_value, if_assign_plus_two(start_value))
+    goal_action = InstantaneousAction("goal_action")
+    goal_action.add_effect(g, mid_value)
+    problem = Problem("interpreted_functions_minimal_chain_of_assignments")
+    problem.add_fluent(start_value)
+    problem.add_fluent(mid_value)
+    problem.add_fluent(g)
+    problem.add_action(start_action)
+    problem.add_action(mid_action)
+    problem.add_action(goal_action)
+    problem.set_initial_value(start_value, 1)
+    problem.set_initial_value(mid_value, 1)
+    problem.set_initial_value(g, 1)
+    problem.add_goal(GE(g, 4))
+
+    ifproblem = TestCase(
+        problem=problem,
+        solvable=True,
+        valid_plans=[
+            up.plans.SequentialPlan([start_action(), mid_action(), goal_action()])
+        ],
+        invalid_plans=[
+            up.plans.SequentialPlan([]),
+        ],
+    )
+    problems["interpreted_functions_minimal_chain_of_assignments"] = ifproblem
+
     return problems
